+++ changelog.md
@@ -72,3 +72,4 @@
 
 ## 0.3.14
 - Added additional files and custom ids to assets
+- Added additional files and custom ids to kits